refactor : Class 분리
Car, Game, Application(Main) 로 분리.
